requirements.txt
streamlit
Pillow
opencv-python-headless
dlib
DeepFace
rembg
numpy